@tarcieri 2023-06-15
Oh hmm, the target features were stabilized sometime after 1.59. Guess I'll have to figure out when.