@tarcieri 2023-06-15
Weird, it wouldn't compile on `aarch64-unknown-linux-gnu` without 10debe5 which causes a performance regression.

Let me play around with it a bit more. It's working fine on `aarch64-apple-darwin` and I would think `#[inline(always)]` would work in place of `target_feature`.