@newpavlov 2023-06-15
>Weird, it wouldn't compile on aarch64-unknown-linux-gnu without https://github.com/RustCrypto/block-ciphers/commit/10debe547fdae5612fdf72fb445ad5ed4904b27e which causes a performance regression.

Huh, it's certainly looks weird. I don't think I've seen such behavior on x86. I would think that asm blocks should be completely opaque for compiler (apart from register allocation stuff). Maybe ask about it on IRLO or in a new Rust repository issue?