@tarcieri 2023-06-15
Yeah, seems like some sort of bug in rustc. You can see the build failure here:

https://github.com/RustCrypto/block-ciphers/actions/runs/5283017067/jobs/9558698657

Curiously it only seems to care about `AESE` and `AESIMC`. It doesn't seem to care about `AESD` and `AESMC`.