@tarcieri 2023-06-15
I'm pretty perplexed by the performance difference of https://github.com/RustCrypto/block-ciphers/commit/10debe547fdae5612fdf72fb445ad5ed4904b27e too... it seems like unrelated code makes the performance greatly suffer

```
$ RUSTFLAGS="--cfg aes_armv8" cargo +nightly bench
   Compiling aes v0.8.2 (/Users/tony/src/RustCrypto/block-ciphers/aes)
    Finished bench [optimized] target(s) in 1.40s
     Running unittests src/lib.rs (/Users/tony/src/RustCrypto/block-ciphers/target/release/deps/aes-eb55ebb5f7ca4ca0)

running 4 tests
test armv8::test_expand::aes128_key_expansion ... ignored
test armv8::test_expand::aes128_key_expansion_inv ... ignored
test armv8::test_expand::aes192_key_expansion ... ignored
test armv8::test_expand::aes256_key_expansion ... ignored

test result: ok. 0 passed; 0 failed; 4 ignored; 0 measured; 0 filtered out; finished in 0.00s

     Running benches/mod.rs (/Users/tony/src/RustCrypto/block-ciphers/target/release/deps/mod-6305260e7c7483e7)

running 15 tests
test aes128_decrypt_block  ... bench:       1,060 ns/iter (+/- 9) = 15456 MB/s
test aes128_decrypt_blocks ... bench:       1,027 ns/iter (+/- 14) = 15953 MB/s
test aes128_encrypt_block  ... bench:       2,698 ns/iter (+/- 43) = 6072 MB/s
test aes128_encrypt_blocks ... bench:       3,606 ns/iter (+/- 170) = 4543 MB/s
test aes128_new            ... bench:         104 ns/iter (+/- 4)
test aes192_decrypt_block  ... bench:       1,447 ns/iter (+/- 14) = 11322 MB/s
test aes192_decrypt_blocks ... bench:       1,273 ns/iter (+/- 11) = 12870 MB/s
test aes192_encrypt_block  ... bench:       3,032 ns/iter (+/- 93) = 5403 MB/s
test aes192_encrypt_blocks ... bench:       3,607 ns/iter (+/- 172) = 4542 MB/s
test aes192_new            ... bench:         103 ns/iter (+/- 3)
test aes256_decrypt_block  ... bench:       1,755 ns/iter (+/- 85) = 9335 MB/s
test aes256_decrypt_blocks ... bench:       1,571 ns/iter (+/- 28) = 10429 MB/s
test aes256_encrypt_block  ... bench:       3,417 ns/iter (+/- 22) = 4794 MB/s
test aes256_encrypt_blocks ... bench:       4,148 ns/iter (+/- 268) = 3949 MB/s
test aes256_new            ... bench:         124 ns/iter (+/- 1)

test result: ok. 0 passed; 0 failed; 0 ignored; 15 measured; 0 filtered out; finished in 22.36s
```