@newpavlov 2023-06-15
My guess is that it fails to inline the "intrinsics", despite having `#[target_feature(enable = "aes")]` on their users. It could be worth to explore code generation in godbolt to find what exactly causes codegen degradation.