@tarcieri 2023-06-16
@newpavlov what's really weird is that the performance of `vaesdq_u8_and_vaesimcq_u8` is massively degraded, but it doesn't call either `vaeseq_u8` or `vaesimcq_u8`, it has its own separate block of assembly.

Replacing `#[inline(always)]` with `#[target_feature(enable = "aes")]` is causing a performance degradation on seemingly unrelated code. Perhaps you're right that it's breaking inlining somehow, but the functions are only related by their contents (i.e. the same ASM instructions), not a direct call path.