@tarcieri 2023-06-16
I'm also having trouble extracting a minimal repro. Just `encrypt1`/`encrypt8`/`decrypt1`/`decrypt8` and the intrinsics seem to be fine, so I guess it's something with the key expansion.