@tarcieri 2023-06-16
I opened a rustc issue here: https://github.com/rust-lang/rust/issues/112709