@tarcieri 2023-06-16
Okay, it seems there were some callers in the key expansion code which weren't properly annotated with `target_feature(enable = "aes")` which were the culprit: https://github.com/RustCrypto/block-ciphers/pull/365/commits/c7fe62edcdb121b709af1689dda1214f435007af

The OS-specific discrepancies were due to `target_feature=+aes` being enabled on certain targets but not others, with the error occurring where it wasn't enabled.

With that new commit we're green on all targets, and I've confirmed performance is not impacted, so this is ready for review.