@tarcieri 2023-06-16
@newpavlov hmm, perhaps we should use `#[inline]` + `#[target_feature]` instead of `#[inline(always)]` to avoid these hard-to-debug errors: https://github.com/rust-lang/stdarch/pull/306