@tarcieri 2023-06-17
Hmm, interesting, on an M2 Max the "parallel" versions seem to offer no speedup and in fact perform worse:

```
test aes128_decrypt_block  ... bench:         838 ns/iter (+/- 24) = 19551 MB/s
test aes128_decrypt_blocks ... bench:         851 ns/iter (+/- 9) = 19252 MB/s
test aes128_encrypt_block  ... bench:         845 ns/iter (+/- 46) = 19389 MB/s
test aes128_encrypt_blocks ... bench:         853 ns/iter (+/- 26) = 19207 MB/s
test aes128_new            ... bench:          64 ns/iter (+/- 4)
test aes192_decrypt_block  ... bench:         984 ns/iter (+/- 27) = 16650 MB/s
test aes192_decrypt_blocks ... bench:       1,004 ns/iter (+/- 46) = 16318 MB/s
test aes192_encrypt_block  ... bench:       1,002 ns/iter (+/- 65) = 16351 MB/s
test aes192_encrypt_blocks ... bench:       1,041 ns/iter (+/- 90) = 15738 MB/s
test aes192_new            ... bench:          62 ns/iter (+/- 3)
test aes256_decrypt_block  ... bench:       1,164 ns/iter (+/- 31) = 14075 MB/s
test aes256_decrypt_blocks ... bench:       1,349 ns/iter (+/- 105) = 12145 MB/s
test aes256_encrypt_block  ... bench:       1,180 ns/iter (+/- 186) = 13884 MB/s
test aes256_encrypt_blocks ... bench:       1,340 ns/iter (+/- 69) = 12226 MB/s
test aes256_new            ... bench:          84 ns/iter (+/- 5)
```