@newpavlov 2023-06-17
@tarcieri 
Such micro-benchmarks are quite finicky, especially on sub-microsecond scales. The difference is within the measurement error, so I would say the performance is effectively the same.

Also note that the best number of blocks processed in parallel is processor-dependent. It depends not only on latency/throughput of used instructions, but also on number of available registers. So even within the same target arch it may vary. 8 blocks is the optimal number for mainstream x86 processors, while on certain older x86 CPUs 6 blocks would perform better. So on M1/M2 a different number could produce better results, or maybe we even don't need explicit parallel processing for it, since these CPUs are famously wide with deep reordering buffers.