@tarcieri 2023-06-17
@newpavlov for AES-256 in particular it's outside the noise threshold, where the noise floor would be around 1,271 ns and the non-parallel version is around 1,180 ns. It's also fairly consistent and reproducible.

I added the parallelism because there was a noticeable benefit on earlier M1 CPUs however, which are much more widespread.

Anyway, just an observation.