@newpavlov 2023-06-17
Maybe try to experiment with different numbers of parallel blocks? It would be interesting to see a plot with such data for M1 and M2.

Also, maybe there are Apple recommendations for implementing AES similar to the Intel ones?